change!: move `fetch::agent|indicate_end_of_interaction` to crate root
These are generally useful in all interactions, including push.
